chore: renamed siwx types folder
