+++ packages/common/src/types/siwx/config.ts
