+++ packages/common/src/types/siwx/verifier.ts
